chore: activate docs check for PRs
